requirements.txt
absl-py==1.2.0
appdirs==1.4.4
asgiref==3.5.2
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
Django==4.1
flatbuffers==1.12
fleep==1.0.1
fonttools==4.37.1
gast==0.4.0
grpcio==1.48.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
imutils==0.5.4
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
librosa==0.9.2
mediapipe==0.8.10.1
numba==0.56.2
numpy==1.23.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python-headless
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
tensorflow-cpu
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
django-heroku
